ref : NoStockMessage
